[RISC-V][MC] Support RVY AUIPC instruction
For RVY, the AUIPC instruction is mode-dependent, writing back a full
capability (X*_Y) register in capability mode and a plain X register
otherwise. This introduces the NullReg RegisterByHwMode to support the
lpad InstAlias that uses a fixed register and therefore we have to use a
register that varies by HwMode in the pattern.

This was made possible by the previous commits that introduced
`RegisterByHwMode`. An alternative would have been to try to support static
resolution of the `RegClassByHwMode` based on the predicates, but that
requires duplicating all `InstAlias`es, which would be fine in the single
case here, but we also need this for JALR with lots of `InstAlias`es.

Pull Request: https://github.com/llvm/llvm-project/pull/177074