[TableGen] Slightly improve error location for a fatal error
I was hitting this error and the error location was pointing to the
register class definition instead of the incorrect InstAlias. Pass in
the InstAlias location to make it easier to debug.

Happens with `def : InstAlias<"foo", (Inst X0)>``, where `Inst` takes
a RegClassByHwMode operand that is not necessarily satisfied by
register X0. Not test yet in this commit since I plan to improve this
error message in a follow-up.