[RISC-V][MC] Support Zaamo, Zalrsc, and Zabha in RVY mode
This is fairly straightforward, just need to introduce a new
BasePtrZeroOffset (instead of GPRMemZeroOffset) that has a
mode-dependent register operand. We could have also updated the existing
GPRMemZeroOffset since this behaviour is correct for all extensions
when RVY is present, but I want to do this one extension at a time to
ensure that we have RVY test coverage for each extension.